Change handleNextQuestion() placement
